+++ stockai/yahoo/stock.py
@@ -6,25 +6,37 @@ class Stock(Base):
         super(Stock, self).__init__(symbol)
         self.refresh()
 
-    # Summary Profile
     def get_summary_profile(self):
-        return self.data_set['summaryProfile']
+      """Return summary profile of a security"""
+      return self.data_set['summaryProfile']
 
-    # Financial Data
+    def get_info(self):
+      """Return detail information of a security"""
+      return self.data_set['price']
+
     def get_price(self):
-        return self.data_set['financialData']['currentPrice']['fmt']
+      """Return price of a security"""
+      return self.data_set['financialData']['currentPrice']['fmt']
 
     def get_currency(self):
-        return self.data_set['financialData']['financialCurrency']
+      """Return currency of a security"""
+      return self.data_set['financialData']['financialCurrency']
 
-    # Historical Prices
     def get_historical_prices(self, start_date, end_date):
-        if (date_to_timestamp(start_date) > date_to_timestamp(end_date)):
-            raise ValueError('Please check the order of start date and end date')
+      """Return historical prices of a security
+      Parameters
+      ----------
+      start_date: str
+        Start date
+      end_date: str
+        End date
+      """
+      if (date_to_timestamp(start_date) > date_to_timestamp(end_date)):
+          raise ValueError('Please check the order of start date and end date')
 
-        return self.get_historical(date_to_timestamp(start_date), date_to_timestamp(end_date))
+      return self.get_historical(date_to_timestamp(start_date), date_to_timestamp(end_date))
 
-    # Get all prices
     def get_all_prices(self):
+      """Return all historical prices starting from 01/01/2000"""
       return self.get_all_historical()