+++ tests/yahoo.py
@@ -17,6 +17,9 @@ class TestStock(TestCase):
       # print(self.td.get_summary_profile())
       self.assertEqual(self.td.get_summary_profile()['city'], 'Toronto')
 
+    def test_td_get_info(self):
+      self.assertEqual(self.td.get_info()['symbol'], 'TD.TO')
+
     def test_td_financial_data(self):
       float(self.td.get_price())
       self.assertEqual(self.td.get_currency(), 'CAD')